Optimize point in ring
Since ring is an interface, the compiler cannot prove that the
closure created in ringContainsPoint is not needed after the
call to ring.Search completes. This causes the closure to be
allocated on the heap, reducing performance due to excessive
heap allocations. In tile38, this is on the hot path for WITHIN
commands on complex polygons.